requirements.txt
pyzmq
pyserial
pylibftdi
zaber.serial
git+https://github.com/kshalm/zmqhelpers

